US 0007 - Unfollow
Adicionando a funcionalidade de Unfollow. Presente também a verificação se o usuário existe e verificando se o usuário já segue a pessoa que ele quer seguir.

+++ src/main/java/com/desafio_spring/demo/model/user/Client.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class Client extends User{
 
-    public Client(Long id, String name, TypeUser type, List<FollowingRelationships> followingRelationships) {
+    public Client(Integer id, String name, TypeUser type, List<FollowingRelationships> followingRelationships) {
         super(id, name, TypeUser.CLIENT, followingRelationships);
     }
 }
